refactor private methods to separate files
