+++ backprop_helpers.py
@@ -1,87 +1,55 @@
 import numpy as np
 
-def linear_activation_forward(A_prev, W, b, activation):
+def back_prop(AL, Y, caches, lambd):
     """
-    Implement the forward propagation for the LINEAR->ACTIVATION layer
-
-    Parameters
-    ----------
-    A_prev : numpy array
-        activations from previous layer (or input data): 
-        (size of previous layer, number of examples)
-    W : numpy array
-        weights matrix of shape (current layer size, previous layer size)
-    b : numpy array
-        Bias vector of shape (current layer size, 1)
-    activation : str
-        Activation to be used in this layer, "sigmoid" or "relu"
-
-    Returns
-    -------
-    A : numpy array
-        Output of the activation function 
-    cache : tuple
-        "linear_cache" and "activation_cache" stored for computing 
-        the backward pass efficiently
-    """
-    
-    Z = np.dot(W, A_prev) + b
-    linear_cache = (A_prev, W, b)
-
-    if activation == "sigmoid":
-        A, activation_cache = sigmoid(Z)
-    
-    elif activation == "relu":
-        A, activation_cache = relu(Z)
-    
-    assert(Z.shape == (W.shape[0], A_prev.shape[1]))
-    assert (A.shape == (W.shape[0], A_prev.shape[1]))
-    cache = (linear_cache, activation_cache)
-
-    return A, cache
-
-def sigmoid(Z):
-    """
-    Implements the sigmoid activation in numpy
+    Implement the backward propagation for the [LINEAR->RELU] * (L-1) -> LINEAR -> SIGMOID group
     
     Parameters
     ----------
-    Z : numpy array
+    AL : numpy array
+        Output of the forward propagation (forward_prop())
+    Y : numpy array
+        True "label" vector (containing 0 if non-cat, 1 if cat)
+    caches : list
+        list of caches containing:
+            every cache of linear_activation_forward() with "relu" (it's caches[l], for l in range(L-1) i.e l = 0...L-2)
+            the cache of linear_activation_forward() with "sigmoid" (it's caches[L-1])
     
     Returns
     -------
-    A : numpy array
-    cache : numpy array
+    gradients : dict
+        A dictionary with the gradients
+            gradients["dA" + str(l)] = ... 
+            gradients["dW" + str(l)] = ...
+            gradients["db" + str(l)] = ... 
     """
+    gradients = {}
+    L = len(caches) # the number of layers
+    Y = Y.reshape(AL.shape) # after this line, Y is the same shape as AL
     
-    A = 1 / (1 + np.exp(-Z))
-    cache = Z
+    # Initializing the backpropagation
+    dAL = - (np.divide(Y, AL) - np.divide(1 - Y, 1 - AL))
     
-    return A, cache
+    # Lth layer (SIGMOID -> LINEAR) gradients. Inputs: "dAL, current_cache". Outputs: "gradients["dAL-1"], gradients["dWL"], gradients["dbL"]
+    current_cache = caches[L-1]
+    gradients["dA" + str(L-1)], gradients["dW" + str(L)], gradients["db" + str(L)] = linear_activation_backward(dAL,
+                                                                                                    current_cache,
+                                                                                                    "sigmoid",
+                                                                                                    lambd)        
+    # Loop from l=L-2 to l=0
+    for l in reversed(range(L-1)):
+        # lth layer: (RELU -> LINEAR) gradients.
+        # Inputs: "gradients["dA" + str(l + 1)], current_cache". Outputs: "gradients["dA" + str(l)] , gradients["dW" + str(l + 1)] , gradients["db" + str(l + 1)] 
+        current_cache = caches[l]
+        dA_prev_temp, dW_temp, db_temp = linear_activation_backward(gradients["dA" + str(l + 1)],
+                                                                    current_cache,
+                                                                    "relu",
+                                                                    lambd)
+        gradients["dA" + str(l)] = dA_prev_temp
+        gradients["dW" + str(l + 1)] = dW_temp
+        gradients["db" + str(l + 1)] = db_temp
 
-def relu(Z):
-    """
-    Implement the RELU function.
-
-    Parameters
-    ----------
-    Z : numpy array
-        Output of the linear layer, of any shape
-
-    Returns
-    -------
-    A : numpy array
-        Post-activation parameter, of the same shape as Z
-    cache : dict
-        a python dictionary containing "A" ; stored for computing the backward pass efficiently
-    """
-    
-    A = np.maximum(0,Z)
-    
-    assert(A.shape == Z.shape)
-    
-    cache = Z 
-    return A, cache
+    return gradients
 
 def linear_activation_backward(dA, cache, activation, lambd):
     """
