+++ loss_function_helpers.py
