+++ multilayer_perceptron.py
@@ -1,9 +1,11 @@
 import math
 import numpy as np
 import matplotlib.pyplot as plt
-from propagation_helpers import linear_activation_forward, linear_activation_backward
-from loss_function import compute_cost, regularized_cost
-from sklearn import preprocessing
+from initialization_helpers import initialize_parameters, initialize_adam_parameters, initialize_minibatches
+from fwdprop_helpers import forward_prop
+from loss_function_helpers import compute_cost, regularized_cost
+from backprop_helpers import back_prop
+from parameter_helpers import gd_update, adam_update
 
 class MultiLayerNN:
     """
@@ -32,305 +34,6 @@ class MultiLayerNN:
         self.Y = Y
         self.parameters = {}
         self.final_cost = 0
-    
-    """
-    INITIALIZATION
-    """
-
-    def initialize_parameters(self, layer_dimensions):
-        """
-        Parameters
-        ----------
-        layer_dimensions : list
-            Dimensions of each layer in the neural network
-        
-        Returns
-        -------
-        parameters: python dictionary
-            Parameters "W1", "b1", ..., "WL", "bL":
-                        Wi : weight matrix of shape (layer_dimensions[i],
-                                                     layer_dimensions[i - 1])
-                        bi : bias vector of shape (layer_dimensions[i], 1)
-        """
-        parameters = {}
-        L = len(layer_dimensions)
-
-        for i in range(1, L):
-            parameters['W' + str(i)] = np.random.randn(layer_dimensions[i],
-                layer_dimensions[i - 1]) * np.sqrt(2 / layer_dimensions[i - 1])
-            parameters['b' + str(i)] = np.zeros((layer_dimensions[i], 1))
-        
-            assert(parameters['W' + str(i)].shape == (layer_dimensions[i], 
-                                                      layer_dimensions[i - 1]))
-            assert(parameters['b' + str(i)].shape == (layer_dimensions[i], 1))
-        
-        return parameters
-    
-    def initialize_adam_parameters(self) :
-        """
-        Initializes v and s as two python dictionaries with:
-            - keys: "dW1", "db1", ..., "dWL", "dbL" 
-            - values: numpy arrays of zeros of the same shape as the 
-                      corresponding gradients/parameters.
-        
-        Returns
-        ------- 
-        v : dict
-            Exponentially weighted average of the gradients.
-                v["dW" + str(l)] = ...
-                v["db" + str(l)] = ...
-        s : dict
-            Exponentially weighted average of the squared gradients.
-                s["dW" + str(l)] = ...
-                s["db" + str(l)] = ...
-
-        """
-        
-        L = len(self.parameters) // 2 # number of layers in the neural networks
-        v = {}
-        s = {}
-        
-        # Initialize v, s. Input: "parameters". Outputs: "v, s".
-        for l in range(L):
-            v["dW" + str(l+1)] = np.zeros(self.parameters["W" + str(l+1)].shape)
-            v["db" + str(l+1)] = np.zeros(self.parameters["b" + str(l+1)].shape)
-            s["dW" + str(l+1)] = np.zeros(self.parameters["W" + str(l+1)].shape)
-            s["db" + str(l+1)] = np.zeros(self.parameters["b" + str(l+1)].shape)
-        
-        return v, s
-
-    def initialize_minibatches(self, minibatch_size = 64, seed = 0):
-        """
-        Creates a list of random minibatches from (X, Y)
-        
-        Parameters
-        ----------
-        minibatch_size : int
-            Size of the mini-batches, integer
-        
-        Returns
-        -------
-        minibatches : list
-            List of synchronous (minibatch_X, minibatch_Y)
-        """
-        
-        np.random.seed(seed)
-        m = self.X.shape[1] # number of training examples
-        minibatches = []
-            
-        # Shuffle (X, Y)
-        permutation = list(np.random.permutation(m))
-        shuffled_X = self.X[:, permutation]
-        shuffled_Y = self.Y[:, permutation].reshape((1, m))
-
-        # Partition (shuffled_X, shuffled_Y). Minus the end case.
-        num_complete_minibatches = math.floor(m / minibatch_size)
-        for k in range(0, num_complete_minibatches):
-            minibatch_X = shuffled_X[:, k * minibatch_size : (k + 1) * minibatch_size]
-            minibatch_Y = shuffled_Y[:, k * minibatch_size : (k + 1) * minibatch_size]
-            minibatch = (minibatch_X, minibatch_Y)
-            minibatches.append(minibatch)
-        
-        # Handling the end case (last mini-batch < minibatch_size)
-        if m % minibatch_size != 0:
-            final_batch_size = m - minibatch_size * math.floor(m / minibatch_size)
-            minibatch_X = shuffled_X[:, num_complete_minibatches * minibatch_size : num_complete_minibatches * minibatch_size + final_batch_size]
-            minibatch_Y = shuffled_Y[:, num_complete_minibatches * minibatch_size : num_complete_minibatches * minibatch_size + final_batch_size]
-            minibatch = (minibatch_X, minibatch_Y)
-            minibatches.append(minibatch)
-        
-        return minibatches
-
-    """
-    FORWARD PROP
-    """
-
-    def forward_prop(self, X, parameters):
-        """
-        Implement forward propagation for the [LINEAR->RELU]*(L-1)->LINEAR->SIGMOID computation
-        
-        Parameters
-        ----------
-        X : numpy array
-            Output of initialize_parameters() of shape 
-            (input size, number of examples)
-        
-        Returns
-        -------
-        AL : numpy array
-            last post-activation value
-        caches : list
-            Every cache of linear_activation_forward() 
-            (there are L-1 of them, indexed from 0 to L-1)
-        """
-        caches = []
-        A = X
-        L = len(parameters) // 2 
-        
-        # Implement [LINEAR -> RELU]*(L-1). Add "cache" to the "caches" list.
-        for l in range(1, L):
-            A_prev = A 
-            A, cache = linear_activation_forward(A_prev,
-                                                 parameters['W' + str(l)],
-                                                 parameters['b' + str(l)],
-                                                 activation = "relu")
-            caches.append(cache)
-        
-        # Implement LINEAR -> SIGMOID. Add "cache" to the "caches" list.
-        AL, cache = linear_activation_forward(A,
-                                              parameters['W' + str(L)],
-                                              parameters['b' + str(L)],
-                                              activation = "sigmoid")
-        caches.append(cache)
-        
-        assert(AL.shape == (1, X.shape[1]))
-                
-        return AL, caches
-    
-    """
-    BACKPROP 
-    """
-    
-    def back_prop(self, AL, Y, caches, lambd):
-        """
-        Implement the backward propagation for the [LINEAR->RELU] * (L-1) -> LINEAR -> SIGMOID group
-        
-        Parameters
-        ----------
-        AL : numpy array
-            Output of the forward propagation (forward_prop())
-        Y : numpy array
-            True "label" vector (containing 0 if non-cat, 1 if cat)
-        caches : list
-            list of caches containing:
-                every cache of linear_activation_forward() with "relu" (it's caches[l], for l in range(L-1) i.e l = 0...L-2)
-                the cache of linear_activation_forward() with "sigmoid" (it's caches[L-1])
-        
-        Returns
-        -------
-        gradients : dict
-            A dictionary with the gradients
-                gradients["dA" + str(l)] = ... 
-                gradients["dW" + str(l)] = ...
-                gradients["db" + str(l)] = ... 
-        """
-        gradients = {}
-        L = len(caches) # the number of layers
-        Y = Y.reshape(AL.shape) # after this line, Y is the same shape as AL
-        
-        # Initializing the backpropagation
-        dAL = - (np.divide(Y, AL) - np.divide(1 - Y, 1 - AL))
-        
-        # Lth layer (SIGMOID -> LINEAR) gradients. Inputs: "dAL, current_cache". Outputs: "gradients["dAL-1"], gradients["dWL"], gradients["dbL"]
-        current_cache = caches[L-1]
-        gradients["dA" + str(L-1)], gradients["dW" + str(L)], gradients["db" + str(L)] = linear_activation_backward(dAL,
-                                                                                                        current_cache,
-                                                                                                        "sigmoid",
-                                                                                                        lambd)        
-        # Loop from l=L-2 to l=0
-        for l in reversed(range(L-1)):
-            # lth layer: (RELU -> LINEAR) gradients.
-            # Inputs: "gradients["dA" + str(l + 1)], current_cache". Outputs: "gradients["dA" + str(l)] , gradients["dW" + str(l + 1)] , gradients["db" + str(l + 1)] 
-            current_cache = caches[l]
-            dA_prev_temp, dW_temp, db_temp = linear_activation_backward(gradients["dA" + str(l + 1)],
-                                                                        current_cache,
-                                                                        "relu",
-                                                                        lambd)
-            gradients["dA" + str(l)] = dA_prev_temp
-            gradients["dW" + str(l + 1)] = dW_temp
-            gradients["db" + str(l + 1)] = db_temp
-
-        return gradients
-    
-    """
-    UPDATE PARAMETERS
-    """
-
-    def gd_update(self, parameters, gradients, learning_rate):
-        """
-        Update parameters using gradient descent
-        
-        Parameters
-        ----------
-        parameters : dict
-        gradients : dict
-            Gradients output by back_prop()
-        
-        Returns
-        -------
-        parameters : dict
-            Updated parameters:
-                parameters["W" + str(l)] = ... 
-                parameters["b" + str(l)] = ...
-        """
-        
-        L = len(parameters) // 2 
-
-        # Update rule for each parameter
-        for l in range(L):
-            parameters["W" + str(l+1)] -= learning_rate * gradients["dW" + str(l+1)]
-            parameters["b" + str(l+1)] -= learning_rate * gradients["db" + str(l+1)]
-        return parameters
-    
-    def adam_update(self, parameters, gradients, v, s, adam_counter, learning_rate = 0.01,
-                    beta1 = 0.9, beta2 = 0.999,  epsilon = 1e-8):
-        """
-        Update parameters using Adam
-        
-        Parameters
-        ----------
-        parameters : dict
-            parameters['W' + str(l)] = Wl
-            parameters['b' + str(l)] = bl
-        gradients : dict
-            gradients['dW' + str(l)] = dWl
-            gradients['db' + str(l)] = dbl
-        v : dict
-            Adam variable, moving average of the first gradient
-        s : dict
-            Adam variable, moving average of the squared gradient
-        learning_rate : float
-        beta1 : float
-            Exponential decay hyperparameter for the first moment estimates 
-        beta2 : float
-            Exponential decay hyperparameter for the second moment estimates 
-        epsilon : float
-            hyperparameter preventing division by zero in Adam updates
-
-        Returns
-        -------
-        parameters : dict
-        v : dict
-        s : dict
-        """
-        
-        L = len(parameters) // 2  # number of layers in the neural networks
-        v_new = {}          # Initializing first moment estimate
-        s_new = {}          # Initializing second moment estimate
-        
-        # Perform Adam update on all parameters
-        for l in range(L):
-            # Moving average of the gradients. Inputs: "v, gradients, beta1". Output: "v".
-            v["dW" + str(l+1)] = beta1 * v["dW" + str(l+1)] + (1 - beta1) * gradients["dW" + str(l+1)]
-            v["db" + str(l+1)] = beta1 * v["db" + str(l+1)] + (1 - beta1) * gradients["db" + str(l+1)]
-
-            # Compute bias-corrected first moment estimate. Inputs: "v, beta1, adam_counter". Output: "v_new".
-            v_new["dW" + str(l+1)] = v["dW" + str(l+1)] / (1 - np.power(beta1, adam_counter))
-            v_new["db" + str(l+1)] = v["db" + str(l+1)] / (1 - np.power(beta1, adam_counter))
-
-            # Moving average of the squared gradients. Inputs: "s, gradients, beta2". Output: "s".
-            s["dW" + str(l+1)] = beta2 * s["dW" + str(l+1)] + (1 - beta2) * np.power(gradients["dW" + str(l+1)], 2)
-            s["db" + str(l+1)] = beta2 * s["db" + str(l+1)] + (1 - beta2) * np.power(gradients["db" + str(l+1)], 2)
-
-            # Compute bias-corrected second raw moment estimate. Inputs: "s, beta2, adam_counter". Output: "s_new".
-            s_new["dW" + str(l+1)] = s["dW" + str(l+1)] / (1 - np.power(beta2, adam_counter))
-            s_new["db" + str(l+1)] = s["db" + str(l+1)] / (1 - np.power(beta2, adam_counter))
-
-            # Update parameters. Inputs: "parameters, learning_rate, v_new, s_new, epsilon". Output: "parameters".
-            parameters["W" + str(l+1)] = parameters["W" + str(l+1)] - learning_rate * v_new["dW" + str(l+1)] / (np.sqrt(s_new["dW" + str(l+1)]) + epsilon)
-            parameters["b" + str(l+1)] = parameters["b" + str(l+1)] - learning_rate * v_new["db" + str(l+1)] / (np.sqrt(s_new["db" + str(l+1)]) + epsilon)
-
-        return parameters, v, s
 
     """
     MODEL
@@ -388,13 +91,13 @@ class MultiLayerNN:
         m = self.X.shape[1]       # number of training examples
         
         # Initialize parameters
-        self.parameters = self.initialize_parameters(layer_dimensions)
+        self.parameters = initialize_parameters(layer_dimensions)
 
         # Initialize the optimizer
         if optimizer == "gd":
             pass # no optimizer used for vanilla gradient descent
         elif optimizer == "adam":
-            v, s = self.initialize_adam_parameters()
+            v, s = initialize_adam_parameters(self.parameters)
         else:
             raise ValueError("The only supported optimizer modes are 'gd' and 'adam'.")
         
@@ -408,7 +111,7 @@ class MultiLayerNN:
                 # Define the random minibatches. We increment the seed to 
                 # reshuffle differently the dataset after each epoch
                 seed = seed + 1
-                minibatches = self.initialize_minibatches(minibatch_size, seed)
+                minibatches = initialize_minibatches(self.X, self.Y, minibatch_size, seed)
 
                 for minibatch in minibatches:
 
@@ -416,28 +119,28 @@ class MultiLayerNN:
                     (minibatch_X, minibatch_Y) = minibatch
 
                     # Forward propagation: [LINEAR -> RELU]*(L-1) -> LINEAR -> SIGMOID.
-                    AL, caches = self.forward_prop(minibatch_X, self.parameters)
+                    AL, caches = forward_prop(minibatch_X, self.parameters)
 
                     # Compute cost and add to the cost total
                     cost_total += regularized_cost(AL, minibatch_Y, self.parameters, lambd)
 
                     # Backward propagation
-                    gradients = self.back_prop(AL, minibatch_Y, caches, lambd)
+                    gradients = back_prop(AL, minibatch_Y, caches, lambd)
 
                     # Update parameters
                     if optimizer == "gd":
-                        self.parameters = self.gd_update(self.parameters, gradients,
+                        self.parameters = gd_update(self.parameters, gradients,
                                                          learning_rate)
                     elif optimizer == "adam":
                         adam_counter += 1 # Adam counter
-                        self.parameters, v, s = self.adam_update(self.parameters,
+                        self.parameters, v, s = adam_update(self.parameters,
                             gradients, v, s, adam_counter, learning_rate,
                             beta1, beta2, epsilon)
             else:
-                AL, caches = self.forward_prop(self.X, self.parameters)
+                AL, caches = forward_prop(self.X, self.parameters)
                 cost_total += regularized_cost(AL, self.Y, self.parameters, lambd)
-                gradients = self.back_prop(AL, self.Y, caches, lambd)
-                self.parameters = self.gd_update(self.parameters, gradients, learning_rate)
+                gradients = back_prop(AL, self.Y, caches, lambd)
+                self.parameters = gd_update(self.parameters, gradients, learning_rate)
 
             # Print the cost every 1000 epoch
             cost_avg = cost_total / m
@@ -481,7 +184,7 @@ class MultiLayerNN:
         p = np.zeros((1,m))
         
         # Forward propagation
-        probabilities = self.forward_prop(X, self.parameters)[0]
+        probabilities = forward_prop(X, self.parameters)[0]
 
         # convert probabilities to 0/1 predictions
         for i in range(0, probabilities.shape[1]):
